Set stack size to 4KB in the default linker script
The linker script used in Aery32 framework has been adapted from the AVR32 GNU
Toolchain (the default one that compiler uses if none was given). In the
default version of this linker script, the heap and stack sections aren't
defined by default and should be defined by the user. This commit fixes
that by setting stack to 4KB. In addition major clean up has been made for
the file.

In future it might be a good idea to define PHRDS sections too.